Removed RollButton to reduce coupling between architecture.
- Rather than the back-end logic of rolling dice being relegated to a
specific button, the logic is triggered by the board facade. This
allows the architecture to maintain the rule that only facades are able
to access back-end logic.
- Button components in the control menu (i.e. Roll Dice and End Turn)
are now in the board facade in order to allow the Catan class to purely
focus on control handling (how it's planned is that the controls check
buttons within facades and if the applicable button is clicked, to then
perform the appropriate action by accessing the facades then finally
the back-end).