Fixed resource bank bug, Minor UI improvements
- Resource counts would not respect the maximum amount of resources
that the board has. The bugfix implemented fixes this.
- Roll Dice number and Tile tokens now adjust the number location more
accurately.
- Trading now properly trades resources with both the bank and the
active player (had to make relationship between player and board
bidirectional to implement this cleanly).
- Development Card UI components are now hidden since Development Cards
will not be utilized for the upcoming release.
- Victory points required to win is now 6 and the quantity of each
resources that the board starts off with is 30 to make up for the fact
that the upcoming release only contains core game mechanics.